+++ src/back/erpetaplication/src/main/java/com/erpet/erpetaplication/service/IServiceCategoria.java
@@ -8,13 +8,13 @@ public interface IServiceCategoria
 {
 
 	public Categoria cadastrarCategoria(Categoria categoria);
-	public Categoria editarCategoria(Long id, Categoria categoria);
-	public Categoria excluirCategoria(Long idCategoria);
+	public Categoria editarCategoria(int id, Categoria categoria);
+	public Categoria excluirCategoria(int idCategoria);
 	public List<Categoria> buscarTodos();
 	public List<Categoria> buscarTodosNaoExcluidos();
 	public List<Categoria> buscarPorNome(String nome);
 	public List<Categoria> buscarPorDescricao(String descricao);
-	public Categoria buscarPorId(Long id);
+	public Categoria buscarPorId(int id);
 	//TODO buscar por nome do produto
     //TODO filtrar por categoria não excluidos
